requirements.txt
langchain==0.2.13
python-dotenv==1.0.1
streamlit==1.22.0
tiktoken==0.7.0
faiss-cpu==1.8.0
protobuf~=3.19.0
google-generativeai==0.7.2
InstructorEmbedding==1.0.1
sentence_transformers==2.2.2
langchain-community==0.2.12
langchain-google-genai==1.0.8